backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg[binary]==3.2.3
python-dateutil==2.8.2
reportlab==4.0.9
